+++ domain_alias/src/Entity/DomainAlias.php
@@ -4,6 +4,7 @@ namespace Drupal\domain_alias\Entity;
 
 use Drupal\domain_alias\DomainAliasInterface;
 use Drupal\Core\Config\Entity\ConfigEntityBase;
+use Drupal\Core\Entity\EntityStorageInterface;
 
 /**
  * Defines the domain alias entity.